Updates for Standard Edition
Instance name is modified because SQLEXPRESS is used as an instance name for SQL Server. In Standard or Enterprise installs, SQL Server will default to MSSQLSERVER. Any instance name declared by the user will have MSSQL$ appeneded to the front as the service name.
Agent name needs to be declared because if you use the SQL Agent, you need to restart both services as the Agent is dependent on the SQL Service
If you have declared an agent account it will restart both the agent service and the sql service. If not only restart the sql service
SQL Server requires a reboot to complete your install